X86/vmsi: Hook guest MSI injection
In two situations, hypervisor injects a msi to a hvm guest. One is when qemu
sends a request to hypervisor through XEN_DMOP_inject_msi. The other is when a
physical interrupt arrives and it has been binded to a msi interrupt of guest.
The binding relationship is setup and updated by qemu.

When interrupt remapping enabled, the binding implementation needs to be
extended because some useful fields which are useless if interrupt remapping
is disabled are discarded in the process. So, we add a new type guest
interrupt struct hvm_gmsi_ir_info and a new type flag
HVM_IRQ_DPCI_GUEST_MSI_IR which means the binding relationship is between
physical interrupts and guest remappable msi.

Signed-off-by: Chao Gao <chao.gao@intel.com>
Signed-off-by: Lan Tianyu <tianyu.lan@intel.com>